requirements.txt
fastapi
uvicorn[standard]
python-multipart
gTTS
speechrecognition
supabase
python-dotenv
aiofiles
requests
pydub
httpx
